requirements.txt
Flask
Flask-Admin
Flask-BabelEx
Flask-Bootstrap
Flask-Login
Flask-Migrate
Flask-Moment
Flask-QRcode
Flask-Security
Flask-SQLAlchemy
email-validator
itsdangerous
Jinja2
python-telegram-handler
pylatexenc
cityhash
qrcode